spark_llm_dataquality/requirements.txt
tensorflow
tensorflow-datasets
datasets
google-cloud-bigquery
google-cloud-storage